+++ .gitignore
@@ -1,6 +1,6 @@
 node_modules
 standard
-
+dist
 
 # vim swap files
 .*.swp
