+++ package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "scripts": {
     "install": "scripts/install-git-hooks",
-    "postinstall": "ln -s node_modules/.bin/standard ."
+    "postinstall": "ln -s node_modules/.bin/standard .",
+    "build": "webpack --config webpack.config.js"
   },
   "repository": {
     "type": "git",
@@ -18,8 +19,18 @@
   },
   "homepage": "https://github.com/afcidk/HackMDir#readme",
   "devDependencies": {
+    "@babel/core": "^7.4.3",
+    "@babel/preset-env": "^7.4.3",
+    "babel-loader": "^8.0.5",
+    "babel-polyfill": "^6.26.0",
+    "css-loader": "^2.1.1",
+    "node-sass": "^4.11.0",
+    "sass-loader": "^7.1.0",
     "standard": "^12.0.1",
-    "webpack-chrome-extension-reloader": "^1.2.2"
+    "style-loader": "^0.23.1",
+    "webpack": "^4.30.0",
+    "webpack-chrome-extension-reloader": "^1.2.2",
+    "webpack-cli": "^3.3.1"
   },
   "dependencies": {
     "pre-commit": "^1.2.2",
