+++ src/store/index.js
@@ -1,0 +1,5 @@
+const ui = require('./modules/ui')
+
+module.exports = {
+  ui
+}
