+++ src/store/modules/ui.js
@@ -1,0 +1,7 @@
+const state = {
+  displayRoot: false
+}
+
+module.exports = {
+  state
+}
